Search Function Added

+++ view/articles.view.php
@@ -24,7 +24,13 @@
                 </div>
                 <div class="card-body">
                     <?php printAlert(); ?>
-                    <table class="table table-striped table-responsive-sm">
+                    <?php if (empty($search_results) && isset($search_query)): ?>
+                        <div class='alert alert-danger' role='alert'>Sorry, no results for <?= $search_query ?> </div>
+                        
+                    <?php elseif (isset($search_query)): ?>
+                        <div class='alert alert-success' role='alert'>Search Results for <?= $search_query ?> </div>
+                    <?php endif ?>
+                    <table class="table table-striped table-responsive">
                             <thead>
                                 <tr class="text-center">
                                     <th scope="col">ID</th>
@@ -41,31 +47,60 @@
                             </thead>
                             
                             <tbody>
-                                <?php foreach ($articles as $key => $article): ?>
-                                    <tr>
-                                        <th scope="row"><?= $key + 1 ?></th>
-                                        <td><a href="/articles?id=<?= $article['id'] ?>" class="btn btn-transparent"><img class="icon" src="<?= BASE_URL ?>/assets/icons/icons8-surprise-64.png" alt="Preview Icon"></a></td>
-                                        <td><?= removeStyles($article['title']) ?></td>
-                                        <td><?= compressText(removeStyles($article['description'])) ?></td>
-                                        <td><?= compressText(removeStyles($article['contents'])) ?></td>
-                                        <?php $category = filterFromDBData($categories, $article['category_id']); ?>
-                                        <td><?= $category['title'] ?></td>
-                                        <?php $user = filterFromDBData($users, $article['user_id']); ?>
-                                        <td class="text-nowrap"><?= $user['name'] ?></td>
-                                        <td><?= date('d M Y', strtotime($article['created_at'])) ?></td>
-                                        <td><?= date('d M Y', strtotime($article['updated_at'])) ?></td>
-                                        <td>
-                                            <div class="d-flex justify-content-around justify-content-md-center">
-                                                    <a href="<?= BASE_URL ?>/article/manipulate/edit?id=<?= $article['id']?>" class="btn p-1 mr-2  btn-transparent edit-btn">
-                                                        <img src="<?= BASE_URL ?>/assets/icons/icons8-edit.svg" alt="Edit Icon" class="icon">
-                                                    </a>
-                                                    <a href="<?= BASE_URL ?>/article/manipulate/delete?id=<?= $article['id']?>&csrf_token=<?= $csrf_token ?>" class="btn p-1 btn-transparent delete-btn" onclick="return confirm('Are you sure want to Delete?')">
-                                                        <img src="<?= BASE_URL ?>/assets/icons/icons8-trash.svg" alt="Delete Icon" class="icon">
-                                                    </a>
-                                            </div>
-                                        </td>
-                                    </tr>
-                                <?php endforeach ?>
+                                <?php if (!empty($search_results)): ?>
+
+                                    <?php foreach ($search_results as $key => $article): ?>
+                                        <tr>
+                                            <th scope="row"><?= $key + 1 ?></th>
+                                            <td><a href="/articles?id=<?= $article['id'] ?>" class="btn btn-transparent"><img class="icon" src="<?= BASE_URL ?>/assets/icons/icons8-surprise-64.png" alt="Preview Icon"></a></td>
+                                            <td><?= removeStyles($article['title']) ?></td>
+                                            <td><?= compressText(removeStyles($article['description'])) ?></td>
+                                            <td><?= compressText(removeStyles($article['contents'])) ?></td>
+                                            
+                                            <td><?= $article['category'] ?></td>
+                                            
+                                            <td class="text-nowrap"><?= $article['username'] ?></td>
+                                            <td><?= date('d M Y', strtotime($article['created_at'])) ?></td>
+                                            <td><?= date('d M Y', strtotime($article['updated_at'])) ?></td>
+                                            <td>
+                                                <div class="d-flex justify-content-around justify-content-md-center">
+                                                        <a href="<?= BASE_URL ?>/article/manipulate/edit?id=<?= $article['id']?>" class="btn p-1 mr-2  btn-transparent edit-btn">
+                                                            <img src="<?= BASE_URL ?>/assets/icons/icons8-edit.svg" alt="Edit Icon" class="icon">
+                                                        </a>
+                                                        <a href="<?= BASE_URL ?>/article/manipulate/delete?id=<?= $article['id']?>&csrf_token=<?= $csrf_token ?>" class="btn p-1 btn-transparent delete-btn" onclick="return confirm('Are you sure want to Delete?')">
+                                                            <img src="<?= BASE_URL ?>/assets/icons/icons8-trash.svg" alt="Delete Icon" class="icon">
+                                                        </a>
+                                                </div>
+                                            </td>
+                                        </tr>
+                                    <?php endforeach ?>
+                                <?php else: ?>
+                                    <?php foreach ($articles as $key => $article): ?>
+                                        <tr>
+                                            <th scope="row"><?= $key + 1 ?></th>
+                                            <td><a href="/articles?id=<?= $article['id'] ?>" class="btn btn-transparent"><img class="icon" src="<?= BASE_URL ?>/assets/icons/icons8-surprise-64.png" alt="Preview Icon"></a></td>
+                                            <td><?= removeStyles($article['title']) ?></td>
+                                            <td><?= compressText(removeStyles($article['description'])) ?></td>
+                                            <td><?= compressText(removeStyles($article['contents'])) ?></td>
+                                            <?php $category = filterFromDBData($categories, $article['category_id']); ?>
+                                            <td><?= $category['title'] ?></td>
+                                            <?php $user = filterFromDBData($users, $article['user_id']); ?>
+                                            <td class="text-nowrap"><?= $user['name'] ?></td>
+                                            <td><?= date('d M Y', strtotime($article['created_at'])) ?></td>
+                                            <td><?= date('d M Y', strtotime($article['updated_at'])) ?></td>
+                                            <td>
+                                                <div class="d-flex justify-content-around justify-content-md-center">
+                                                        <a href="<?= BASE_URL ?>/article/manipulate/edit?id=<?= $article['id']?>" class="btn p-1 mr-2  btn-transparent edit-btn">
+                                                            <img src="<?= BASE_URL ?>/assets/icons/icons8-edit.svg" alt="Edit Icon" class="icon">
+                                                        </a>
+                                                        <a href="<?= BASE_URL ?>/article/manipulate/delete?id=<?= $article['id']?>&csrf_token=<?= $csrf_token ?>" class="btn p-1 btn-transparent delete-btn" onclick="return confirm('Are you sure want to Delete?')">
+                                                            <img src="<?= BASE_URL ?>/assets/icons/icons8-trash.svg" alt="Delete Icon" class="icon">
+                                                        </a>
+                                                </div>
+                                            </td>
+                                        </tr>
+                                    <?php endforeach ?>
+                                <?php endif ?>
                             </tbody>
                         </table>
                 </div>
